*more work in progress *lots of internal gui control work *added isfs (nand) filesystem (read only) *compile fix for devkitPPC r24

+++ trunk/source/ImageOperations/ImageViewer.cpp
@@ -504,7 +504,7 @@ bool ImageViewer::LoadImage(int index, bool silent)
 		return false;
 	}
 
-	u64 filesize;
+	u32 filesize;
 	u8 * file = NULL;
 
 	int ret = -1;
